refactor: change default icon for projects
