+++ apps/hub/src/domains/group/views/group-projects.tsx
@@ -2,7 +2,6 @@ import { BillingPlanBadge } from "@/domains/project/components/billing/billing-p
 import { ProjectTableActions } from "@/domains/project/components/project-table-actions";
 import { groupProjectsQueryOptions } from "@/domains/project/queries";
 import {
-  AssetImage,
   Button,
   Card,
   CardContent,
@@ -47,7 +46,6 @@ export function GroupProjects({ groupId }: GroupProjectsProps) {
           <Table>
             <TableHeader>
               <TableRow>
-                <TableHead w="16" />
                 <TableHead>Name</TableHead>
                 <TableHead w="16" />
               </TableRow>
@@ -60,30 +58,22 @@ export function GroupProjects({ groupId }: GroupProjectsProps) {
                   </TableCell>
                 </TableRow>
               ) : null}
-              {data.projects.map((game) => (
+              {data.projects.map((project) => (
                 <TableRow
-                  key={game.gameId}
+                  key={project.gameId}
                   isClickable
                   onClick={() => {
                     navigate({
                       to: "/projects/$projectId",
-                      params: { projectId: game.gameId },
+                      params: { projectId: project.gameId },
                     });
                   }}
                 >
-                  <TableCell w="16">
-                    <AssetImage
-                      src={game.logoUrl || "/games/blank/blankgame.svg"}
-                      alt={`${game.displayName} logo`}
-                      width={64}
-                      height={64}
-                    />
-                  </TableCell>
                   <TableCell>
                     <Text asChild>
                       <Flex items="center" gap="2">
-                        {game.displayName}{" "}
-                        <BillingPlanBadge projectId={game.gameId} />
+                        {project.displayName}{" "}
+                        <BillingPlanBadge projectId={project.gameId} />
                       </Flex>
                     </Text>
                   </TableCell>
