+++ apps/hub/src/domains/project/components/group-project-select.tsx
@@ -1,6 +1,5 @@
 import { groupProjectsQueryOptions } from "@/domains/project/queries";
 import {
-  AssetImage,
   Flex,
   Select,
   SelectContent,
@@ -60,14 +59,7 @@ export function GroupProjectSelect({
         ) : null}
         {data.projects.map((project) => (
           <SelectItem key={project.gameId} value={project.gameId}>
-            <Flex gap="2" items="center">
-              <AssetImage
-                src={project.logoUrl || "/games/blank/blankgame.svg"}
-                className="mx-auto size-5 object-contain"
-                alt="Project logo"
-              />
-              {project.displayName}
-            </Flex>
+            {project.displayName}
           </SelectItem>
         ))}
       </SelectContent>
