+++ packages/icons/manifest.json
@@ -8060,6 +8060,9 @@
   "@awesome.me/kit-63db24046b/icons/kit/custom": {
     "icons": [
       { "icon": "faActors", "aliases": [] },
+      { "icon": "faGb", "aliases": [] },
+      { "icon": "faProject", "aliases": [] },
+      { "icon": "faRivet", "aliases": [] },
       { "icon": "faSelect", "aliases": [] }
     ]
   }
